% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_it_shred.R
\name{make_it_shred}
\alias{make_it_shred}
\title{Add 1980s guitar lord as a plot watermark}
\usage{
make_it_shred(
  shred_lord = c("Vai", "Satriani", "Gilbert", "Malmsteen", "Hammett", "Sambora",
    "Friedman", "Becker")
)
}
\description{
Add 1980s guitar lord as a plot watermark
}
