% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shredR.R
\docType{data}
\name{shredR_palette}
\alias{shredR_palette}
\title{General colour palette imports as hex codes}
\format{
An object of class \code{character} of length 7.
}
\usage{
shredR_palette
}
\description{
General colour palette imports as hex codes
}
\keyword{datasets}
