% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shredR.R
\name{theme_shredR}
\alias{theme_shredR}
\title{Style general plot features according to the shredR theme}
\usage{
theme_shredR(grids = FALSE)
}
\description{
Style general plot features according to the shredR theme
}
